requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
filterpy>=1.4.5
scipy>=1.11.0
ultralytics>=8.0.0  # YOLO desteği için
scikit-learn>=1.3.0  # İleri analiz için 